Support int64 literal goldens

+++ tests/golden/outputs/glsl/conversion-float-to-int.glsl
@@ -19,12 +19,12 @@ uint test_f16_to_u32(float16_t f) {
     return uint(f);
 }
 
-int test_f16_to_i64(float16_t f) {
-    return int(f);
+int64_t test_f16_to_i64(float16_t f) {
+    return int64_t(f);
 }
 
-uint test_f16_to_u64(float16_t f) {
-    return uint(f);
+uint64_t test_f16_to_u64(float16_t f) {
+    return uint64_t(f);
 }
 
 int test_f32_to_i32(float f) {
@@ -35,12 +35,12 @@ uint test_f32_to_u32(float f) {
     return uint(f);
 }
 
-int test_f32_to_i64(float f) {
-    return int(f);
+int64_t test_f32_to_i64(float f) {
+    return int64_t(f);
 }
 
-uint test_f32_to_u64(float f) {
-    return uint(f);
+uint64_t test_f32_to_u64(float f) {
+    return uint64_t(f);
 }
 
 int test_f64_to_i32(double f) {
@@ -51,12 +51,12 @@ uint test_f64_to_u32(double f) {
     return uint(f);
 }
 
-int test_f64_to_i64(double f) {
-    return int(f);
+int64_t test_f64_to_i64(double f) {
+    return int64_t(f);
 }
 
-uint test_f64_to_u64(double f) {
-    return uint(f);
+uint64_t test_f64_to_u64(double f) {
+    return uint64_t(f);
 }
 
 ivec2 test_f16_to_i32_vec(vec2 f) {
@@ -112,30 +112,30 @@ void test_const_eval() {
     int max_f16_to_i32 = int(global_1);
     uint min_f16_to_u32 = uint(global_0);
     uint max_f16_to_u32 = uint(global_1);
-    int min_f16_to_i64 = int(global_0);
-    int max_f16_to_i64 = int(global_1);
-    uint min_f16_to_u64 = uint(global_0);
-    uint max_f16_to_u64 = uint(global_1);
+    int64_t min_f16_to_i64 = int64_t(global_0);
+    int64_t max_f16_to_i64 = int64_t(global_1);
+    uint64_t min_f16_to_u64 = uint64_t(global_0);
+    uint64_t max_f16_to_u64 = uint64_t(global_1);
     int min_f32_to_i32 = int(global_2);
     int max_f32_to_i32 = int(global_3);
     uint min_f32_to_u32 = uint(global_2);
     uint max_f32_to_u32 = uint(global_3);
-    int min_f32_to_i64 = int(global_2);
-    int max_f32_to_i64 = int(global_3);
-    uint min_f32_to_u64 = uint(global_2);
-    uint max_f32_to_u64 = uint(global_3);
-    int min_f64_to_i64 = int(global_4);
-    int max_f64_to_i64 = int(global_5);
-    uint min_f64_to_u64 = uint(global_4);
-    uint max_f64_to_u64 = uint(global_5);
+    int64_t min_f32_to_i64 = int64_t(global_2);
+    int64_t max_f32_to_i64 = int64_t(global_3);
+    uint64_t min_f32_to_u64 = uint64_t(global_2);
+    uint64_t max_f32_to_u64 = uint64_t(global_3);
+    int64_t min_f64_to_i64 = int64_t(global_4);
+    int64_t max_f64_to_i64 = int64_t(global_5);
+    uint64_t min_f64_to_u64 = uint64_t(global_4);
+    uint64_t max_f64_to_u64 = uint64_t(global_5);
     int min_abstract_float_to_i32 = int(global_6);
     int max_abstract_float_to_i32 = int(global_7);
     uint min_abstract_float_to_u32 = uint(global_6);
     uint max_abstract_float_to_u32 = uint(global_7);
-    int min_abstract_float_to_i64 = int(global_6);
-    int max_abstract_float_to_i64 = int(global_7);
-    uint min_abstract_float_to_u64 = uint(global_6);
-    uint max_abstract_float_to_u64 = uint(global_7);
+    int64_t min_abstract_float_to_i64 = int64_t(global_6);
+    int64_t max_abstract_float_to_i64 = int64_t(global_7);
+    uint64_t min_abstract_float_to_u64 = uint64_t(global_6);
+    uint64_t max_abstract_float_to_u64 = uint64_t(global_7);
 }
 
 void wgsl_main() {
